Change type of gainAudio from int16_t to int8_t
The gainAudio may be in the range 1 to 100. So we can use int8_t type to reduce the computation load.

+++ SimpleAudioFilter07/audioIO.cpp
@@ -84,7 +84,7 @@ int           passInput=0;
 uint8_t filterMode=0; 
 uint8_t nrMode=0;
 int16_t outSample=0;
-int16_t gainAudio =1;
+int8_t gainAudio =1;
 
 // Check if need to boost the audio 
 // For safe reasons the value will be
